refactor(adk): move ralph-loop example from turn-loop to agent directory
The Ralph Loop pattern no longer uses TurnLoop — it drives the agent
directly via Runner in a simple external for-loop. This makes the
validate-then-retry logic external and explicit, matching the feedback
that completion verification should not live inside TurnLoop callbacks.

Also fixes seed files to use space indentation so the edit_file tool
matches content correctly (tab indentation was ambiguous due to
read_file's line-number formatting).
